# backend/docs/swagger.yaml
definitions:
  handlers.BinPackingResponse:
    properties:
      ai_recommendation:
        type: string
      bin_dimensions:
        example:
        - 600
        - 400
        - 400
        items:
          type: number
        type: array
      execution_time:
        example: 12.4ms
        type: string
      fill_percentage:
        example: 84.5
        type: number
      placements:
        items:
          $ref: '#/definitions/handlers.PackedPlacement'
        type: array
    type: object
  handlers.BoxStyleResponse:
    properties:
      category:
        example: Shipping
        type: string
      code:
        example: FEFCO 0201
        type: string
      description:
        example: Standard meeting flap box
        type: string
      name:
        example: Regular Slotted Container
        type: string
    type: object
  handlers.ChatMessage:
    properties:
      content:
        example: Pack 2 boxes of 200x150x100mm into a standard container.
        type: string
      role:
        example: user
        type: string
    type: object
  handlers.MaterialResponse:
    properties:
      max_weight_kg:
        example: 10
        type: number
      profile:
        example: E-Flute
        type: string
      thickness_mm:
        example: 1.5
        type: number
    type: object
  handlers.PackRequestPayload:
    properties:
      messages:
        items:
          $ref: '#/definitions/handlers.ChatMessage'
        type: array
      retrieval_k:
        example: 3
        type: integer
      temperature:
        example: 0.7
        type: number
      top_k:
        example: 40
        type: integer
    type: object
  handlers.PackedPlacement:
    properties:
      color:
        example: '#2563eb'
        type: string
      dim:
        example:
        - 200
        - 150
        - 100
        items:
          type: number
        type: array
      item_id:
        example: item-101
        type: string
      pos:
        example:
        - 0
        - 0
        - 0
        items:
          type: number
        type: array
    type: object
  main.PackagingRecommendation:
    properties:
      box_style:
        example: FEFCO 0201
        type: string
      dieline_svg:
        type: string
      status:
        example: success
        type: string
    type: object
  main.ProductInput:
    properties:
      height:
        example: 150
        type: number
      length:
        example: 300
        type: number
      weight:
        example: 2.5
        type: number
      width:
        example: 200
        type: number
    type: object
info:
  contact: {}
paths:
  /api/v1/materials:
    get:
      description: Retrieves cardboard flute profiles and load limits directly from
        SQLite.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.MaterialResponse'
            type: array
        "500":
          description: Database error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List Cardboard Flute Profiles
      tags:
      - Catalog
  /api/v1/pack:
    post:
      consumes:
      - application/json
      description: Accepts chat messages, retrieval parameters, and LLM hyperparameters
        to return optimized 3D coordinate placements.
      parameters:
      - description: Packing agent request payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PackRequestPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BinPackingResponse'
        "400":
          description: Invalid request payload
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: LLM or DB error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Run 3D Bin Packing Optimization Agent
      tags:
      - Spatial Engine
  /api/v1/recommend:
    post:
      consumes:
      - application/json
      description: Accepts product specs, checks Postgres for clearance rules, and
        queries Gemini @ Temp 0.0
      parameters:
      - description: Product Specification Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.ProductInput'
      produces:
      - application/json
      responses:
        "200":
          description: Recommended FEFCO box and SVG layout
          schema:
            $ref: '#/definitions/main.PackagingRecommendation'
        "400":
          description: Invalid input error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal processing error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Calculate optimal packaging box and dieline
      tags:
      - Packaging
  /api/v1/styles:
    get:
      description: Fetches available FEFCO standard container specifications directly
        from SQLite.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.BoxStyleResponse'
            type: array
        "500":
          description: Database error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List FEFCO Structural Box Styles
      tags:
      - Catalog
swagger: "2.0"